Fixing preboot issues + moving all types to typedef file
