+++ .env.dev
@@ -1,3 +1,4 @@
 ENABLE_CORS=true
 PORT=3000
 CONNECTION_TIMEOUT=-1
+MAX_CONCURRENT_SESSIONS=1
