+++ src/hooks.ts
@@ -5,9 +5,11 @@ import { Socket } from 'net';
 import * as path from 'path';
 import { Page } from 'puppeteer';
 
-import { IWebdriverStartHTTP } from './browserless';
-import { IBrowser } from './chrome-helper';
-import * as util from './utils';
+import {
+  IBrowser,
+  IWebdriverStartHTTP,
+  IHTTPRequest,
+} from './types';
 
 const beforeHookPath = path.join(__dirname, '..', 'external', 'before.js');
 const afterHookPath = path.join(__dirname, '..', 'external', 'after.js');
@@ -16,7 +18,7 @@ const pageSetupPath = path.join(__dirname, '..', 'external', 'page.js');
 const externalRoutesPath = path.join(__dirname, '..', 'external', 'routes.js');
 
 export const beforeRequest: (args: {
-  req: util.IHTTPRequest,
+  req: IHTTPRequest,
   res?: http.ServerResponse,
   socket?: Socket,
 }) => boolean =
@@ -25,7 +27,7 @@ export const beforeRequest: (args: {
     () => true;
 
 export const afterRequest: (args: {
-  req: util.IHTTPRequest | IWebdriverStartHTTP,
+  req: IHTTPRequest | IWebdriverStartHTTP,
   start: number,
   status: string,
 }) => boolean = fs.existsSync(afterHookPath) ?
