--- src/models/options.interface.ts
@@ -1,39 +1,0 @@
-import { Feature } from '../features';
-
-export interface IBrowserlessOptions
-  extends IBrowserlessServerConfiguration,
-    IChromeServiceConfiguration {}
-
-interface IBrowserlessServerConfiguration {
-  host: string | undefined;
-  port: number;
-  token: string | null;
-  rejectAlertURL: string | null;
-  queuedAlertURL: string | null;
-  timeoutAlertURL: string | null;
-  errorAlertURL: string | null;
-  healthFailureURL: string | null;
-  metricsJSONPath: string | null;
-  exitOnHealthFailure: boolean;
-  workspaceDir: string;
-  disabledFeatures: Feature[];
-  enableAPIGet: boolean;
-}
-
-export interface IChromeServiceConfiguration {
-  connectionTimeout: number;
-  maxConcurrentSessions: number;
-  maxQueueLength: number;
-  prebootChrome: boolean;
-  demoMode: boolean;
-  functionExternals: string[];
-  functionEnableIncognitoMode: boolean;
-  functionBuiltIns: string[];
-  maxMemory: number;
-  maxCPU: number;
-  keepAlive: boolean;
-  chromeRefreshTime: number;
-  enableCors: boolean;
-  singleRun: boolean;
-  token: string | null;
-}
