--- src/models/sandbox.interface.ts
@@ -1,19 +1,0 @@
-import { LaunchOptions } from 'puppeteer';
-
-export interface ISandboxOpts {
-  builtin: string[];
-  external: boolean | string[];
-  root: string;
-}
-export interface IConfig {
-  code: string;
-  timeout: number;
-  opts?: LaunchOptions;
-  sandboxOpts: ISandboxOpts;
-}
-
-export interface IMessage {
-  event: string;
-  context?: any;
-  error?: string;
-}
