+++ src/tests/integration/http.spec.ts
@@ -1,6 +1,6 @@
 import fetch from 'node-fetch';
 import { BrowserlessServer } from '../../browserless';
-import { IBrowserlessOptions } from '../../models/options.interface';
+import { IBrowserlessOptions } from '../../types';
 import {
   defaultParams,
   killChrome,
